Move notification.service to shared/notification
